pyproject.toml
[project]
name = "offline-file-converter"
version = "0.2.0"
description = "Offline Windows desktop converter for PDF, Word, and images."
requires-python = ">=3.11"
dependencies = [
  "PySide6>=6.7",
  "Pillow>=10.0",
  "PyMuPDF>=1.24",
  "python-docx>=1.1",
]

[project.scripts]
offline-converter = "offline_converter.__main__:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
ocr = [
  "paddlepaddle==2.6.2",
  "paddleocr==2.10.0",
  "setuptools==68.2.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
